// src/generated/models/PowerPlatformforAdminsModel.ts
export interface Environment {
  // Environment ID field.
  id?: string;
  // Environment type field.
  type?: string;
  // Environment name field.
  name?: string;
  // Environment location field.
  location?: string;
  // Environment properties object.
  properties?: {
    // Environment azureRegionHint field.
    azureRegionHint?: string;
    // Display Name of Environment.
    displayName?: string;
    // Created Time of Environment.
    createdTime?: string;
    createdBy?: Principal_CreatedBy;
    // Environment lastModified date and time.
    lastModifiedTime?: string;
    lastModifiedBy?: Principal_ModifiedBy;
    // Environment provisioningState.
    provisioningState?: string;
    capacity?: {
     capacityType: string;
     actualConsumption: number;
     ratedConsumption: number;
     capacityUnit: string;
     updatedOn: string;
   }[];
    addons?: {
     addonType: string;
     allocated: number;
     addonUnit: string;
   }[];
    // Creation Type of Environment.
    creationType?: string;
    // Sku ID of Environment.
    environmentSku?: string;
    // Type of Environment.
    environmentType?: string;
    // Is Default Environment or not.
    isDefault?: boolean;
    // Environment runtimeEndpoints object.
    runtimeEndpoints?: {
      // BusinessAppPlatform runtime endpoint.
      "microsoft.BusinessAppPlatform"?: string;
      // CommonDataModel runtime endpoint.
      "microsoft.CommonDataModel"?: string;
      // PowerApps runtime endpoint.
      "microsoft.PowerApps"?: string;
      // Flow runtime endpoint.
      "microsoft.Flow"?: string;
    };
    // Environment softDeletedTime.
    softDeletedTime?: string;
    // Environment linkedEnvironmentMetadata object.
    linkedEnvironmentMetadata?: {
      // Environment linkedEnvironmentMetadata type.
      type?: string;
      // Environment linkedEnvironmentMetadata resourceId.
      resourceId?: string;
      // Environment linkedEnvironmentMetadata  friendlyName.
      friendlyName?: string;
      // Environment linkedEnvironmentMetadata uniqueName.
      uniqueName?: string;
      // Environment linkedEnvironmentMetadata domainName.
      domainName?: string;
      // Environment linkedEnvironmentMetadata version.
      version?: string;
      // Environment linkedEnvironmentMetadata instanceUrl.
      instanceUrl?: string;
      // Environment linkedEnvironmentMetadata instanceApiUrl.
      instanceApiUrl?: string;
      // Environment linkedEnvironmentMetadata baseLanguage.
      baseLanguage?: number;
      // Environment linkedEnvironmentMetadata instanceState.
      instanceState?: string;
      // Environment linkedEnvironmentMetadata createdTime.
      createdTime?: string;
      // Environment linkedEnvironmentMetadata modifiedTime.
      modifiedTime?: string;
      // Environment linkedEnvironmentMetadata hostNameSuffix.
      hostNameSuffix?: string;
      // Environment linkedEnvironmentMetadata bapSolutionId.
      bapSolutionId?: string;
      // Environment linkedEnvironmentMetadata creationTemplates.
      creationTemplates?: string[];
      // Environment linkedEnvironmentMetadata managementPackageVersion.
      managementPackageVersion?: string;
    };
  };
}
// ...
export interface Principal_ModifiedBy {
  // Modified by principal object ID.
  id?: string;
  // Modified by principal displayName.
  displayName?: string;
  // Modified by principal.
  email?: string;
  // Modified by principal type.
  type?: string;
  // Modified by principal tenant ID.
  tenantId?: string;
  // Modified by userPrincipalName.
  userPrincipalName?: string;
}

export interface Principal_CreatedBy {
  // Api Policy creator principal object ID.
  id?: string;
  // Api Policy creator principal displayName.
  displayName?: string;
  // Api Policy creator principal email.
  email?: string;
  // Api Policy creator principal type.
  type?: string;
  // Api Policy creator principal tenant ID.
  tenantId?: string;
  // Api Policy creator userPrincipalName.
  userPrincipalName?: string;
}